requirements.txt
spotdl
spotipy
librosa
selenium
ruamel.yaml